# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AgentTurnResponse": {
        "description": "One agent reply, including any actions the user can perform on it.",
        "properties": {
          "agent_response": {
            "title": "Agent Response",
            "type": "string"
          },
          "applied_patch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Patch"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "patch_applied_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patch Applied At"
          },
          "proposed_actions": {
            "items": {
              "type": "string"
            },
            "title": "Proposed Actions",
            "type": "array"
          },
          "selected_context": {
            "$ref": "#/components/schemas/ContextSelection"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "turn_index": {
            "title": "Turn Index",
            "type": "integer"
          },
          "user_prompt": {
            "title": "User Prompt",
            "type": "string"
          }
        },
        "required": [
          "id",
          "session_id",
          "turn_index",
          "user_prompt",
          "selected_context",
          "agent_response",
          "created_at"
        ],
        "title": "AgentTurnResponse",
        "type": "object"
      },
      "ApplyPatchBody": {
        "description": "POST body for ``/sessions/{id}/patches/{turn_id}/apply``.\n\nCurrently empty; reserved so the endpoint can accept driver overrides\n(e.g. dry-run) without a breaking-change later. We previously declared\n``dry_run: bool`` but nothing in the agent service consumed it (P1-B9)\n— the field was published as a public contract through the OpenAPI\nschema with no implementation backing it, which is worse than no\nfield at all. Removed; clients sending an empty body still pass schema\nvalidation.",
        "properties": {},
        "title": "ApplyPatchBody",
        "type": "object"
      },
...
      "ContextSelection": {
        "description": "The set of artifacts the user has selected as relevant for the turn.",
        "properties": {
          "extras": {
            "items": {
              "type": "string"
            },
            "title": "Extras",
            "type": "array"
          },
          "files": {
            "items": {
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "title": "Logs",
            "type": "array"
          },
          "tests": {
            "items": {
              "type": "string"
            },
            "title": "Tests",
            "type": "array"
          }
        },
        "title": "ContextSelection",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MissionDetail": {
        "description": "`GET /missions/{id}` payload.\n\nWithholds ``ideal_solution`` and any hidden-test surface; brief is included\nbecause the workspace needs it.",
        "properties": {
          "brief": {
            "default": "",
            "title": "Brief",
            "type": "string"
          },
          "category": {
            "enum": [
              "auth",
              "testing",
              "security",
              "frontend",
              "api",
              "database",
              "refactoring",
              "agent-safety",
              "review",
              "debugging"
            ],
            "title": "Category",
            "type": "string"
          },
          "difficulty": {
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "title": "Difficulty",
            "type": "string"
          },
          "estimated_minutes": {
            "title": "Estimated Minutes",
            "type": "integer"
          },
          "expected_context_recommended": {
            "items": {
              "type": "string"
            },
            "title": "Expected Context Recommended",
            "type": "array"
          },
          "expected_context_required": {
            "items": {
              "type": "string"
            },
            "title": "Expected Context Required",
            "type": "array"
          },
          "expected_diff_lines_p50": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Diff Lines P50"
          },
          "failure_mode_id": {
            "title": "Failure Mode Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "initial_commit": {
            "title": "Initial Commit",
            "type": "string"
          },
          "language_runtime": {
            "anyOf": [
              {
                "enum": [
                  "node20",
                  "python312"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language Runtime"
          },
          "manifest_sha256": {
            "title": "Manifest Sha256",
            "type": "string"
          },
          "published": {
            "default": true,
            "title": "Published",
            "type": "boolean"
          },
          "repo_pack": {
            "title": "Repo Pack",
            "type": "string"
          },
          "short_description": {
            "default": "",
            "title": "Short Description",
            "type": "string"
          },
          "skills_tested": {
            "items": {
              "type": "string"
            },
            "title": "Skills Tested",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          },
          "visible_tests": {
            "items": {
              "type": "string"
            },
            "title": "Visible Tests",
            "type": "array"
          }
        },
        "required": [
          "id",
          "title",
          "difficulty",
          "category",
          "estimated_minutes",
          "failure_mode_id",
          "repo_pack",
          "initial_commit",
          "manifest_sha256"
        ],
        "title": "MissionDetail",
        "type": "object"
      },
...
      "PatchResult": {
        "properties": {
          "added_lines": {
            "default": 0,
            "title": "Added Lines",
            "type": "integer"
          },
          "applied": {
            "title": "Applied",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "files_changed": {
            "items": {
              "type": "string"
            },
            "title": "Files Changed",
            "type": "array"
          },
          "removed_lines": {
            "default": 0,
            "title": "Removed Lines",
            "type": "integer"
          }
        },
        "required": [
          "applied"
        ],
        "title": "PatchResult",
        "type": "object"
      },
      "PromptBody": {
        "description": "POST body for ``/sessions/{id}/prompts``.",
        "properties": {
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextSelection"
              },
              {
                "type": "null"
              }
            ]
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "PromptBody",
        "type": "object"
      },
...
      "SessionCreate": {
        "properties": {
          "mission_id": {
            "title": "Mission Id",
            "type": "string"
          }
        },
        "required": [
          "mission_id"
        ],
        "title": "SessionCreate",
        "type": "object"
      },
      "SessionDetail": {
        "description": "`GET /sessions/{id}` response — adds workspace bootstrap material.",
        "properties": {
          "agent_turns": {
            "default": 0,
            "title": "Agent Turns",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "current_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Commit"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mission": {
            "$ref": "#/components/schemas/MissionDetail"
          },
          "mission_id": {
            "title": "Mission Id",
            "type": "string"
          },
          "sandbox_driver": {
            "default": "local",
            "enum": [
              "docker",
              "local"
            ],
            "title": "Sandbox Driver",
            "type": "string"
          },
          "sandbox_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Id"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "provisioning",
              "active",
              "submitting",
              "graded",
              "abandoned",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "ws_token": {
            "title": "Ws Token",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "mission_id",
          "status",
          "started_at",
          "mission",
          "ws_token"
        ],
        "title": "SessionDetail",
        "type": "object"
      },
      "SessionRead": {
        "description": "Bare session row — used as the response for `POST /sessions`.",
        "properties": {
          "agent_turns": {
            "default": 0,
            "title": "Agent Turns",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "current_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Commit"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mission_id": {
            "title": "Mission Id",
            "type": "string"
          },
          "sandbox_driver": {
            "default": "local",
            "enum": [
              "docker",
              "local"
            ],
            "title": "Sandbox Driver",
            "type": "string"
          },
          "sandbox_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Id"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "provisioning",
              "active",
              "submitting",
              "graded",
              "abandoned",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "mission_id",
          "status",
          "started_at"
        ],
        "title": "SessionRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAgentDojo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sessions": {
      "post": {
        "operationId": "post_session_api_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "operationId": "get_session_endpoint_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a session",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/api/v1/sessions/{session_id}/patches/{turn_id}/apply": {
      "post": {
        "operationId": "post_apply_patch_api_v1_sessions__session_id__patches__turn_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "turn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Turn Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ApplyPatchBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply the agent patch for a specific turn",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/prompts": {
      "post": {
        "operationId": "post_prompt_api_v1_sessions__session_id__prompts_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a prompt and get an agent response",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}